refactor : 지갑 객체 및 테스트 리팩터링
